+++ m3u8/writer_test.go
@@ -787,7 +787,7 @@ func TestNewMasterPlaylistWithAlternatives(t *testing.T) {
 		Type:       "AUDIO",
 		Name:       "main",
 		Default:    true,
-		Autoselect: "YES",
+		Autoselect: true,
 		Language:   "english",
 	}
 	p, e := NewMediaPlaylist(3, 5)
@@ -805,8 +805,8 @@ func TestNewMasterPlaylistWithAlternatives(t *testing.T) {
 	if m.ver != 4 {
 		t.Fatalf("Expected version 4, actual, %d", m.ver)
 	}
-	expected := `#EXT-X-MEDIA:TYPE=AUDIO,GROUP-ID="audio",NAME="main",DEFAULT=YES,` +
-		`AUTOSELECT=YES,LANGUAGE="english",URI="800/rendition.m3u8"`
+	expected := `#EXT-X-MEDIA:TYPE=AUDIO,GROUP-ID="audio",NAME="main",LANGUAGE="english",DEFAULT=YES,` +
+		`AUTOSELECT=YES,URI="800/rendition.m3u8"`
 	if !strings.Contains(m.String(), expected) {
 		t.Fatalf("Master playlist did not contain: %s\nMaster Playlist:\n%v", expected, m.String())
 	}